Issue-20: Имплементация CashDonation, CommodityDonation, EF Core конфигурация по шаблону Table Per Hierarchy.

+++ DevFactoryZ.CharityCRM/Persistence/ICashDonationRepository.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Интерфейс описывает шаблон репозиторий для денежных пожертвований.
     /// </summary>
-    public interface ICashDonationRepository : IRepository<CashDonation, int>
+    public interface ICashDonationRepository : IRepository<CashDonation, long>
     {
     }
 }
