+++ DevFactoryZ.CharityCRM/Persistence/ICommodityDonationRepository.cs
@@ -5,7 +5,7 @@
     /// Интерфейс описывает шаблон репозиторий для пожертвований в виде предметов 
     /// (одежды, продуктов питания, книг и т.п.).
     /// </summary>
-    public interface ICommodityDonationRepository : IRepository<CommodityDonation, int>
+    public interface ICommodityDonationRepository : IRepository<CommodityDonation, long>
     {
     }
 }
